(#9869) Exclude yum repodata from squid cache
The behavior without this patch applied is that the squid caching proxy
running on modi was caching repomd.xml which contained stale checksums
for the retrieved copy of primary.sqlite.bz2.

This is a cache coherency problem.  To avoid this issue this patch
explicitly excludes all URL's which contain /repodata/ from being cached
by setting a maximum age to 0 seconds.  This should result in all URL's
for yum metadata to be passed through to the upstream server.