Add codemetrics module to gather graphite statistics outside metric_fu.
(For example, cloc and the time it takes to run rake spec.)
Fix some bugs in metric_fu.
Filebucket metrics.rake because the default one in the master doesn't set some options we need.
Monkeypatch ZAML over YAML, because yaml is unacceptably slow and memory-hogging in ruby 1.8.
(This will probably not be needed anymore in ruby 1.9.)