(#9998) Monitor all bacula jobs with nagios
Previously only bacula jobs on .com hosts were being monitored by
nagios.  There was also an assumption by the module that the nagios host
and the bacula host were one and the same.  On our .lan subnet, this is
not the case.  We have only one nagios server running on baal, and we
have a gearman worker that is running on wyrd that does all the actual
checks for internal hosts by plucking them out of a queue.  The internal
bacula server is running on bacula01, so the module assumptions were
wrong.

This fix makes the check_bacula.pl checks happen over nrpe, so the host
doing the checking and the host that is running the bacula server do not
have to be the same.  NRPE was also extended to allow for a nrpe.d/
style configurations to be placed in the directory and add nrpe
commands.  This is done through the use of nagios::nrpe::command define.