Improve OpenVPN module so that it support multiple client connections.
This was needed so that we can use our nagios system to monitor the
status of hosts within different zones.  Specifically, the office and
the DC at the same time.  The nagios monitoring has not been implemented
in this commit.  Also, improvements to dave and freebsd hosts were made
to support various bits.