Basic configuration in txt format for rpm-builder
This commit adds a txt file with basic puppet code for configuration of
rpm-builder.  This was a small module I was using to configure
rpm-builder using puppet apply.  I was unclear on exactly how to roll in
these changes, so I just put them here.  I completely understand if this
doesn't get merged.