(#10735) Add freight module to deploy freight
This adds the freight module to pluto's manifest, and adds the freight module.
It adds rcrowleys apt key so that freight can be installed with apt, and it
lays down the freight.conf for it. It creates a vhost for freight with the
parameters for the freight class. It also makes sure gpg-agent is installed on
the machine. Freight should be ready to use after the module has been applied.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>