(#11255) Make apache2ctl -t not output because of duplicate NameVirtualHosts entry.
When using using vHosts, our templates have included in them the
NameVirtualHosts entry, which should be specified only once per port.  When
using multiple vhosts on a single system that all listen on the same port, this
causes the initscript to spit out some messages "NameVirtualHost *:80 has no
VirtualHosts" because it has already been defined.  This is a particular
annoyance when logrotate's post-rotate script reloads Apapche and then emails
around a note about this reload.