(#11631) Fix DNS at DC1
When DNS was initially deployed at DC1 it was done by deploying the zone files
directly, which was incorrect.  The module was designed to allow DynamicDNS
(ddns) udpate so that DHCP clients get DNS names.  It was intended that the
initialfile and the statics file would be layed down, where only the staticfile
would be updated and there would be an $INCLUDE line from the initial file that
would import the statics file.  The trouble was that the DNS server maintains a
journal of the DNS entries that have been updated dynamically and only
periodically writes them to disk, which means that the initial file is
overwritten with what bind thinks it should be, and the $INCLUDE line is no
longer present in the initialfile.

This causes trouble when the puppet module attempts to write the zone file again
and reload the bind service because the serial in the journal and the serial in
the zone file will now be different, and reloading the service causes conflict
so there was not source of authority.

To get around this, this commit will write the zone file initially, but not
update it.  This is done on the Bind::Zone[] using the initialfile parameter.

The git repo `git@git.puppetlabs.net:puppetlabs-dnszones.git` is cloned and
update with a cronjob, a script is run that outputs the needed nsupdate commands
and pipes the output to `nsupdate`.  This ensures that only the only entries
that actually make it into the DNS server are dynamic.