Make older Ubuntus use the ancient package repos for apt.
On older, deprecated, dying Ubuntus we have to use
http://old-releases.ubuntu.com/ for their apt sources.  This is making
special cases for things that _should_ go away, but on the flip side,
there's no harm to it, and other releases may disappear in the future.

Am not sure if this is worth doing, nor have I tested it at this juncture.