Large nginx rebase ahead. Make nginx not split configs for SSL on vhost.
This allows you (nay, actually may require you to regrettably), to set a
default vhost on nginx. The 'listen' directive requires each host/port
combo has a default, and if you do ssl on that port or ipv6, then it must
be default.

You cannot, somewhat obviously, have multiple defaults per port, but
this is not enforced by the Puppetry and must be _manually_ put in
place by you.

The benefits of this, are ultimately a simplified config on the nginx end,
but perhaps not the puppet end.

Make nginx reload instead of restart on configuration changes. This
prevents configuration mistakes from taking down your webserver.
This should stop it doing a stop/start on debian/umbongo, so we don't have
micro outages when we change things. We should do the same for apache too.

Enable v6 listening on nginx servers which have the ip6address fact.

Add in sslonly in unicorn template, as we have hosts that need it, such as
dashboard and puppetmasters.