(#11922) Purge apt sources that are not managed by puppet
This will ensure that nay apt source on a box is managed by puppet, so that we
can turn to the puppet manifests to get the definitive source of what sources
exist on a box.  If we stop managing a source, such as the deprecation of karmic
that we have recently done in our manifests, then the files will get removed so
we don't get errors when, say, cronjobs run.