robopuller: don't break gmail threading for pull requests!
GMail treats a change of the subject line as an inescapably strong indication
that two emails are not part of the same conversation, even when it has other
hints that they are the same.

Given that, using two different subject lines for "new" and "closed" pull
requests means that we end up with two actions on the same entity represented
as two distinct, unconnected threads in GMail. (Also most other MUAs given the
nature and content of the emails.)

This fixes that by stripping out the "state" detail from the subject; it still
exists in the body, so pretty much anyone can still filter these into
different buckets if they really, really want to.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>