(#11715) Make logrotate use nice and ionice on Linux systems
This is in regard to the Debian wishlist bug here:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=652600

This change manages the logrotate cronjob on Linux systems to use nice and
ionice.  This should help to keep the load down on boxes when doing compression
of large log files in the middle of the night.  It just seems the sane way to do
things.