(#10382) Bacula by default
For each node that we have cared to back up, we have needed to make make a
password, set the director, and add a manifest to the node definition.  This has
been cumbersome because we just want to back up the node, and its a lot of steps
just for backup, which makes bringing a new new in more tiresome.

Now, including role::server gets bacula on the client and configured with a
random password generated on the fly and put into store configs.