(#12046) Make forge deployment use bundler
Previously, in order to deploy the forge a lot of manual gem futzing was
required.  Now that gem dependencies have been specified using bundler
in the forge repo, a lot of the gem mucking about in the forge manifest
can go away and deployment should be more reliable.

For this manifest to truly work, the vcsrepo code needs to be at the
latest version that includes a fix so that you can checkout SHA1s.
Either that or the forge repo needs to be manually moved to the correct
revision before the manifest is run.