(#12047) Improve Bacula Module
The bacula module now supports running the director and the storage daemon on
different systems.  This will be good when we decide to move things to a single
director, but need to have multiple storage daemons.

move storage related items out to its own resources and class
move storage resource for director into its own resource, tag it and export it so it can be realized on the director
use bacula_is_storage and include bacula::storage if its set to yes
export resources between storage and director