Make Sles-builder just include nagios, & let the class sort out the nrpe server.
In a5d517f1 we make role::server just include a bunch of classes and let heira
work it out in the class. sles-builder is it's own beast, so it never got
this, as it doesn't use role::server (it is too weird). So just emulate
that change and include nagios and let the class figure it out.