(#12227) Replace dxul/projects
This commit migrates the last production service running on dxul to narvi.  In
the process, there has been much re-factor.  Most of unicorn and nginx code was
taken from the forge module which was done by Ben, the Sire of Brigton.  There
has so been some collateral damage, bacula and mysql to name a couple.

Add bundler::bundle define for easy bundles
If bacula is included, dump each mysql database and export resources for backup