Add mco and dashboard to the list of repos that metric_fu runs on.
To make this work, a couple of things are rearranged in the metric_fu module.
The metrical gem is now used to run the metrics to remove the dependency
on the rake task (which puppet and facter have, but the other projects do not).
Also, because dashboard is a rails project, some extra metrics are run;
consequently some new gem dependencies are added.

Finally, a timeout is set on the metric_fu Exec, because otherwise it
occasionally ran wild, causing puppet to give up on it but leave puppetdlock,
which is Bad.