(#12664) Backup Burji to S3
Without this commit, the /opt/pm and /opt/downloads directories
are not backed up anywhere.  This commit adds a puppet module to
dists that enables backup of the /opt/pm and /opt/downloads
directories to an s3 bucket via the s3cmd package. Additional
directories can easily be backed up by adding them to the init.pp
file. The backup is set via the schedule resource to occur
weekly between 2 and 4 am. It is a non-destructive backup, so
files that have been removed from burji are not removed from the
backup automatically.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>