(#12664) Abstract s3cmd, set exec timeout to 0
This commit rewrites the s3cmd backup module originally
designed to work with burji to work with any node. It
abstracts s3cmd from burji, and moves the actual backing
up of directories to the burji node. Also in this commit
is the assignment of the title as the default value of the
path to backup for the defined type s3cmd::backup_dir as
well as the setting of the timeout to 0 so that the exec
does not fail on larger backups.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>