Change s3cmd::backup_dir to cron-based resource
This patch modifies the s3cmd::backup_dir defined
type to wrap the cron resource type instead of the
exec type. This will obviate the need to specify
virtually unlimited timeouts, as is the case with
the current exec-based type. It also removes the
need to set up a schedule resource for the
backups.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>