Big ugly wordtodapress/markops branch merge.
This is the thing that makes markops work. It's the sync framework/hack
between web01 and a test VM we have here. It's a different way of doing
wordpress installs, a bunch of ssh hack, and some other things on top.
Remove NameVirtualHost from markops vhost, as the apache module does it
now in ports.conf, so this should stop annoying cron emails.