Updated Icinga/Nagios/Monitoring.
This cleans up a lot of code that was used twice. The goal is to help
the team figure out where to correctly make changes to icinga/nrpe so we
can start adding custom checks. There is also code now to setup
naginator which creates the nagios checks from puppetdb on cron instead
of collecting resources.