Variables in yaml were not getting pulled in. Hard code one thing.
Somehow variables on the director were not getting pulled into the
config. This broke everything. Same goes with the hardcoded line for the
ssl cert in the client.conf.erb. The paths that are in that file have to
do with the path on the director only, so hal calling to
/usr/local/foo/bar was breaking things