Remove notify and target on icinga collected resources.
We have target => and notify => on every nagios collected resource. I'm
going to remove them as they don't really do anything currently since we
just override them when we collect it. It might be a source of slowdown
on frigga. I also made host_escalation a real bool instead of a string.
There are some puppet lint fixes in some of the files that I made
changes to.