This is me getting ready for PuppetConf 2014
This module has existed behind closed doors for many years, operating in
production for all of Puppet Labs backup infrastructure.  Recent
commits on this module made it less reusable for the general public, but
due to the private nature, this has not been an issue.

This work cleans out all references to internal infrastructure and moves
the module to be a better steward of Open Source, aiming for simplicity
and reusability.  Rounding off the sharp edges has mostly been done
here, but there may yet be areas which are rough and could be smoothed
out.  Tests, a license and a readme now exist.  Templates have been
cleaned up, and a working demo has been fleshed out to ensure the module
does what it says it will do.  Manifests have been tidied, and hire
lookups have been adjusted for sanity.

Configurations that were not used have been completely removed, along
with functional pieces that should be implemented elsewhere; this module
now only does backups.  We no longer support mysql as a database
backend, though it would be trivial to add, this module should remain
simple.  The architecture of bacula itself has been described as
somewhat complex, and this module should not add to said complexity,
but help manage it.

In any case, here it is.  We hope that its useful outside of our
department and can perhaps serve your environment as well as it has our
own.