Convert storage export to define
This enables the ability to export a fragment resource from the storage
to the director, but as its own resource type.  This means that we are
able to export with one set of variables, and override on realize when
the directory applies those exported resources, and should enable
interoperability between director and storage platforms.