Update conf_dir paths
This work ensures that there is seperation between the param data
between components of the system in multi-platform environments where
paths may differ.

Without this change, the director needs to be on a system where the
conf_dir path is under /etc.  This work also shifts the exported
resource on the client to its own define to allow the ability to
override the conf_dir path on the director at realization.