Update templates for listen_address processing
Without this, the various components listen_address parameter is not set
correctly when the address is INET6.  This work detects the address
family of the received listen_address and takes the correct action in
the configuration.  This config snippet is sourced by several templates
to keep the logic centralized and out of the way.

Along the way a few whitespace changes to align the '='.