Change default on client and job resources
This work adds a new param called 'default_pool' to the client resource
to allow a client to be directed to a pool for job information.
Defaults to 'Default'.

Also here changes the default behaviour of the jobs.  Jobs no longer use
the 'Default' pool, but instead refer to the client for the
'default_pool'.  The result here should remain the same if no parameters
are set by the user.

The various pools have been set to undef.  This allows the user to
specify a pool for Incremental, Differential, and Full backups on the
job, but no longer will direct them there automatically.  This change
will require more work on the admin, but should ultimately lead to
greater flexibility.