Change - allow customization of Director's messages
This adds a `messages` parameter to the `bacula::director` class that
expects a Hash of Hashes where the each outer Hash represents one
Messages resource for the Director with each inner Hash being the
settings to be used for the Messages resource.

For backwards compatibility, the new parameter receives a default value
from `bacula::params::dir_messages` that matches exactly what
`bacula::director` previously establish.

The easiest way to customize the Director's Messages setup is via Hiera.
For example:

bacula::director::messages:
    Daemon:
        mname:          'Daemon'
        append:         '"/var/log/bacula/bacula-dir.log" = all, !skipped'
        console:        'all, !skipped, !saved'
        mail:           'you@example.com = all, !skipped'
        mailcmd:        '"/usr/sbin/bsmtp -s \"daemon message\" %r"'
    Standard:
        mname:          'Standard'
        append:         '"/var/log/bacula/bacula-dir.log" = all, !skipped'
        console:        'all, !skipped, !saved'
        mail:           'you@example.com = all, !skipped'
        mailcmd:        '"/usr/sbin/bsmtp -s \"%t %e for %n\" %r"'
        operator:       'you@example.com = mount'
        operatorcmd:    '"/usr/bin/mail -s \"Intervention needed for %j\" %r"'