Update the list of tested puppet versions
This has needed an update for a while.  Here we drop puppet 4.2 so that
we may test cleanly with the puppet lookup function, which seems to be
relied upon by the automatic param lookup when using the data_provider.
Also here we just add some recent versions to keep pace with the world.