Do not send the storage daemon messages to syslog by default
bacula-sd is the only service logging to syslog by default in the
bacula module (bacula-dir and bacula-fd only log to a file).  By
default, bacula itself does not log to syslog.

The syslog destination is also not tunable, logging *all* messages at
priority 'err'.

For these reasons, align the default logging configuration for the
storage daemon service with the default logging configuration of the
director and the file daemon.