Refactor how db_type is passed between classes
The lookup() function is used to fetch the value of
$bacula::director::db_type from the bacula::storage class.  If the user
has customized bacula::director by passing an explicit value, it is not
used by the bacula::storage class, leading to inconsistency and
surprise.

Both classes might not be instanciated on the same nodes, so accessing
$bacula::director::db_type from bacula::storage is not an option.

Move the $bacula::director::db_type to $bacula::db_type, and access this
variable from both bacula::director and bacula::storage.

Users who customized the value would experience a hard failure
(parameter bacula::director::db_type does not exist) instead of a
behavior change…  But since only PostgreSQL is supported, it should not
be a problem.

The whole story is about nucking the last lookup() call.