Exclude unneeded files from the buit packages
When building a package using the Puppet Development Kit (default
packaging system when building a package with Puppet >= 6.0.0), a bunch
of useless files are not automatically excluded, resulting in a package
larger than necessary and which is longer to process by r10k.

Add a .pdkignore file listing these unneeded files so they are ignored
when building the package to fix these issues.  The package is now 22KB
instead of 8.7MB.