Manage storage directories after installation
When creating a storage directory, we want to set correct ownership for
it but before the bacula-sd package is installed, the bacula user and
group may not be available, resulting in a catalog application failure.

Explicitly require the storage daemon package to make sure the
user/group are available when the directory is created.

While here, add some basic testing for this class.